+++ src/anemoi/transform/filters/cos_sin_mean_wave_direction.py
@@ -25,7 +25,7 @@ class CosSinWaveDirection(MatchingFieldsFilter):
     """A filter to convert mean wave direction to cos() and sin() and back."""
 
     @matching(
-        match="param",
+        select="param",
         forward=("mean_wave_direction",),
         backward=("cos_mean_wave_direction", "sin_mean_wave_direction"),
     )