+++ src/anemoi/transform/filters/w_to_wz.py
@@ -8,6 +8,8 @@
 # nor does it submit to any jurisdiction.
 
 
+from typing import Iterator
+
 import earthkit.data as ekd
 
 from . import filter_registry
@@ -21,7 +23,7 @@ class VerticalVelocity(MatchingFieldsFilter):
     """
 
     @matching(
-        match="param",
+        select="param",
         forward=("w_component", "temperature", "humidity"),
         backward=("wz_component", "temperature", "humidity"),
     )
@@ -33,6 +35,7 @@ class VerticalVelocity(MatchingFieldsFilter):
         temperature="t",
         humidity="q",
     ):
+
         # wind speed in Pa/s
         self.w_component = w_component
         # wind speed in m/s
@@ -40,9 +43,27 @@ class VerticalVelocity(MatchingFieldsFilter):
         self.temperature = temperature
         self.humidity = humidity
 
-    def forward_transform(self, w_component: ekd.Field, temperature: ekd.Field, humidity: ekd.Field) -> ekd.Field:
-        """Pa/s to m/s
-        This will return the fields that are used but not modified (temperature and humidity)
+    def forward_transform(
+        self,
+        w_component: ekd.Field,
+        temperature: ekd.Field,
+        humidity: ekd.Field,
+    ) -> Iterator[ekd.Field]:
+        """Convert vertical wind speed from Pa/s to m/s.
+
+        Parameters
+        ----------
+        w_component : ekd.Field
+            The vertical wind speed in Pa/s.
+        temperature : ekd.Field
+            The temperature field.
+        humidity : ekd.Field
+            The humidity field.
+
+        Returns
+        -------
+        Iterator[ekd.Field]
+            The vertical wind speed in m/s.
         """
 
         level = float(w_component._metadata.get("levelist", None))
@@ -54,9 +75,24 @@ class VerticalVelocity(MatchingFieldsFilter):
         yield temperature
         yield humidity
 
-    def backward_transform(self, wz_component: ekd.Field, temperature: ekd.Field, humidity: ekd.Field) -> ekd.Field:
-        """m/s to Pa/s
-        This will return the fields that are used but not modified (temperature and humidity)
+    def backward_transform(
+        self, wz_component: ekd.Field, temperature: ekd.Field, humidity: ekd.Field
+    ) -> Iterator[ekd.Field]:
+        """Convert vertical wind speed from m/s to Pa/s.
+
+        Parameters
+        ----------
+        wz_component : ekd.Field
+            The vertical wind speed in m/s.
+        temperature : ekd.Field
+            The temperature field.
+        humidity : ekd.Field
+            The humidity field.
+
+        Returns
+        -------
+        Iterator[ekd.Field]
+            The vertical wind speed in Pa/s.
         """
 
         level = float(wz_component._metadata.get("levelist", None))